+++ mcddns/meta.py
@@ -1,7 +1,0 @@
-NAME = __package__
-DESCRIPTION = "Industrial-strength dynamic DNS client"
-VERSION = "1.1.2"
-AUTHOR = "Mike Conigliaro"
-URL = "http://conigliaro.org"
-EMAIL = "mike@conigliaro.org"
-COPYRIGHT = f"""Copyright (c) 2021 {AUTHOR}"""