+++ tests/conftest.py
@@ -1,22 +1,19 @@
 import os
-import pytest
 import tempfile
 
+import pytest
+
 import mcddns
-import mcddns.meta as meta
 import mcddns.provider as provider
 
 
-test_providers_path = os.path.join(os.path.dirname(__file__), "providers")
+test_providers_path = os.path.join(os.path.dirname(__file__), "provider")
 provider.import_modules(test_providers_path)
 
 
 @pytest.fixture
 def test_state_path():
-    return os.path.join(
-        tempfile.gettempdir(),
-        f"{meta.NAME}.test_state"
-    )
+    return os.path.join(tempfile.gettempdir(), f"{mcddns.META['Name']}.test_state")
 
 
 @pytest.fixture