+++ tests/test_cli.py
@@ -1,9 +1,9 @@
-import mcddns.options as options
+import mcddns.cli as cli
 
 
 def test_parse():
     args = "test.DNS -a test.Address test.example.com"
-    opts = options.parse(args=args.split())
+    opts = cli.parse(args=args.split())
 
     assert opts.address_providers == ["test.Address"]
     assert opts.dns_provider == "test.DNS"